close 7 prod-blockers found by analyzer audit
D4 TLS-flow:
* internal cluster URLs switched from https:// to http:// (auth, server,
  worker, secret, migration jobs). uvicorn listens plain HTTP — NetworkPolicy
  default-deny ingress already isolates traffic inside the namespace. Edge
  TLS still comes from dbos-ingress-tls via the Ingress.
* INTROSPECT_TLS_VERIFY=false in ConfigMaps (was true while the URLs were
  https://, so the first introspect would hand-shake-fail).
* Dropped k8s/92-certificates.yaml entirely — leaf certs for
  auth/logging/server/secret were issued by cert-manager but no Deployment
  ever mounted them. install_cert_manager.sh no longer applies it.
* Redis password no longer expanded into the pod's argv (was visible via
  `kubectl get pod redis -o yaml`); the entrypoint reads it from env.

C4 connection pool:
* DB_POOL_SIZE=5 / DB_MAX_OVERFLOW=10 set per service in ConfigMaps.
  Previous defaults (10/20 × 2 workers × 2 replicas) overshot PostgreSQL
  `max_connections=100` under any modest burst.

A5/D1 secret_service P0s:
* `_lazy_reencrypt_if_needed` now `db.commit()`s after the CAS UPDATE —
  without it every lazy re-encrypt was silently rolled back when the
  session closed, so master-key `--finalize` would have refused forever.
* `account_admin` branches in `credential_service.transfer/recover/delete`
  removed. Per the secrets-scope memory account_admin must not see
  plaintext or move credentials — the branches were also unreachable
  (`require_user_context` 403'd before them).
* `migration_status` regex_match cast changed `::int` → `::bigint` to
  survive a future `v9999999999$` wire-format version.
* `_emit_action_failure` helper wired around 9 sensitive operations
  (create/update/delete/transfer/recover/dept_grant_*/role_acl_*) so
  failure cases reach the audit channel — previously every failed write
  raised before the emit and the SIEM saw nothing.

C1/D3 server_service /internal/migration_status:
* New `ops_router` mounted at `/api/server/v1/internal/migration_status`
  guarded by `require_internal_caller(rotation_runner)` (X-Service-Identity
  + timing-safe bearer compare). Calls existing `secrets_migration_service.
  status()` which already aggregates server_account + ipmi_controller
  ciphertext versions. `rotate_master_key.sh --auto-finalize` was 404'ing
  against this path; now closed.
* `service_api_keys` env (`SERVER_INBOUND_SERVICE_API_KEYS`) parsed
  through pydantic `NoDecode` + kv-list/JSON parser so the inbound map
  with `rotation_runner` reaches the dependency.

Performance carry-overs (also P1 from C4):
* `GET /servers` N+1 on storage load replaced with a single
  `list_for_servers` SQL — was up to 501 SELECTs at limit=500.
* `PASSWORD_REVEAL_RATE_LIMIT=10/min` per (account, identity) on
  `GET /server_accounts/{id}` and `GET /servers/{id}/ipmi`.
* `_count_by_version` moved into a SQL GROUP BY (`substring(col,'^v([0-9]+)\$')`)
  with `::bigint` cast — no longer pulls every ciphertext into Python.
* `INGEST_RATE_LIMIT=1000/minute` in logging ConfigMap (was 100/min; a
  single auth-pod login burst exceeded that and dropped audit events).

C2 disaster recovery (3 new CronJobs + offsite hook):
* `k8s/102-master-keys-backup.yaml` — daily 04:00 UTC, runs
  `backup_master_keys.sh` (now also dumps `dbos-ca-key-pair` to ca/).
* `k8s/103-secret-full-backup.yaml` — weekly Sunday 05:00 UTC,
  `backup_secret_full.sh --no-encryption --yes`. Operator-supplied
  passphrase Secret (`dbos-backup-passphrase`) toggles encryption.
* `k8s/104-pg-restore-drill.yaml` — monthly 1st 06:00 UTC,
  `pg_restore_drill.sh --non-interactive` cycles all 5 databases.
* `k8s/105-offsite-backup.yaml` — empty `OFFSITE_RSYNC_DEST` ConfigMap;
  master-keys + secret-full CronJobs run an rsync sidecar that no-ops
  when the value is empty.
* `restore_master_keys.sh --apply` now also re-applies `ca/*.yaml`.

Tests:
* auth: 1697 passed
* logging: 1659 passed
* server: 1784 passed (+22 — ops endpoint + N+1 + reveal-rate-limit + SQL GROUP BY)
* worker: 2407 passed
* secret: 386 passed (+10 — lazy-commit + 9 failure-emit cases)

+++ dbos_server_service/secret_service/src/services/migration_status_service.py
@@ -18,12 +18,15 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from src.schemas.internal import MigrationStatus
 
 
-# Postgres regexp_match возвращает text[] — берём первую группу и кастим к int
-# одной expression'ой, чтобы caller'у не пришлось парсить строку.
+# Postgres regexp_match возвращает text[] — берём первую группу и кастим к bigint
+# одной expression'ой, чтобы caller'у не пришлось парсить строку. bigint, а не
+# int: ::int (INTEGER, max 2^31-1) переполнится на ciphertext с версией
+# v9999999999$ или выше; bigint держит до 2^63-1. Регулярка ловит любое
+# количество цифр, поэтому ограничения сверху на размер версии нет.
 _VERSION_HISTOGRAM_SQL = text(
     """
     SELECT
-        (regexp_match(secret_encrypted, '^v(\\d+)\\$'))[1]::int AS version,
+        (regexp_match(secret_encrypted, '^v(\\d+)\\$'))[1]::bigint AS version,
         count(*) AS rows
     FROM credentials
     GROUP BY 1
